+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Fixed
 - don't run in-memory shrinking if there is not enough memory available.
+- don't mark the project as in IDE sync mode when no tasks where called and no composite build detected.
 
 
 ## [0.7.0-alpha1] - 2024-02-02